[ENG-1500] Handle TypeScript SDK (axios) security vulnerability (#435)
* save some code that makes it easier to get the request body content type

* Update Axios version to 0.27.2 and add useAxios0_27_2 flag

This commit updates the Axios version to 0.27.2 in multiple files and adds a new flag called useAxios0_27_2. This flag can be used to control whether to use Axios 0.27.2 or the latest version. The useAxios0_27_2 flag is added to the generator configuration, TypeScript generation function, JavaGenerateApiRequestBody, KonfigYaml, and TypeScriptAxiosClientCodegen files. This change allows for publishing the TypeScript SDK without making breaking changes.

* docs(changeset): Adds useAxios0_27_2 to TypeScript SDK configuration

* typescript-splitit-old-axios

* typescript-newscatcher-old-axios

* run newscatcher SDK tests

* typescript-humanloop-old-axios

* typescript-leap-workflows-old-axios

* fix old axios by pinning konfig-axios-fetch-adapter

* ran yarn build

* regenerate typescript integration test sdks